move download project to middlware and add to construct viewer toolbar as an action
